backend/requirements.txt
fastapi
sqlalchemy
uvicorn
python-jose
werkzeug
pydantic
